pyproject.toml
[tool.poetry]
name = "readarr_send_to_kindle"
version = "0.1.0"
description = ""
authors = ["Patricio Serrano <serranomorantepatricio@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
sendgrid = "^6.9.7"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
